refactor(buzz): make RESONANCE_BUZZ available without [resonance_tester]
Buzz is excitation, not measurement, so it should not require an
accelerometer or a [resonance_tester] section. Move RESONANCE_BUZZ /
RESONANCE_BUZZ_SWEEP (and the buzz mask/sinusoid helpers) into a new
always-loaded extras/resonance_buzz.py, force-loaded by motion so the
commands exist unconditionally. resonance_tester no longer registers
them (avoids a duplicate-command clash); it keeps TEST_RESONANCES /
SHAPER_CALIBRATE / MEASURE_AXES_NOISE and its accelerometer requirement.
motion.submit_resonance_buzz now imports the mask helper from the new
module.
